pyproject.toml
[project]
name = "ffmpeg-ng"
version = "0.1.0"
description = ""
authors = [
    {name = "alwaysmavs",email = "beingswu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
]

[tool.poetry]
package-mode = false
packages = [{include = "ffmpeg_ng", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
